* origin/main:
  Install doxygen in Doxygen deploy.
  Do not use self-hosted for Doxygen.
  More conversion clean-up.
  Fix more warnings.
  Remove more warnings.
  Remove size_t / int conversion warnings.
  Clean up log messages for tetwild.
  Add a function to extract the tetwild mesh data.
  Add toml for pip installation.
  Fix position independent code issues.
  Fix geogram related linking issues.
  Add basic python bindings.
  added spacing support, fixed parity bug with mgrid
  Fix TriMesh split rollback issue.

+++ app/remeshing/src/remeshing/UniformRemeshing.h
@@ -59,7 +59,7 @@ public:
     {
         Eigen::Vector3d v1p;
         Eigen::Vector3d v2p;
-        int partition_id;
+        size_t partition_id;
     };
     tbb::enumerable_thread_specific<PositionInfoCache> position_cache;
